requirements.txt
asttokens==3.0.1
audioread==3.1.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
Cython==3.2.1
debugpy==1.8.17
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
fsspec==2025.10.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.22.4
openpyxl==3.1.5
packaging==25.0
pandas==1.5.3
parso==0.8.5
pathlib==1.0.1
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
pybind11==3.0.1
pycparser==2.23
Pygments==2.19.2
pyroomacoustics==0.7.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
resampy==0.4.3
scikit-learn==1.6.1
scipy==1.6.3
six==1.17.0
soundfile==0.11.0
soxr==1.0.0
spyder-kernels==2.4.4
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
zipp==3.23.0
